+++ atomic/atom/guimaker/src/browser/guimaker/utils/handler.js
@@ -91,6 +91,11 @@ export default await (async () => {
               over: {},
               up: {},
             },
+            keyboard: {
+              keydown: {},
+              keypress: {},
+              keyup: {},
+            },
           }
         );
       },