+++ Changelog.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add new feature which is provide `htmlengine` object and `convtrigger` function to the workflow shared.The file changer involve both webnodejs/browser/preload.html and deskelectronjs/browser/preload.html. Completed on 2024-09-04
 - Apply trycatch error handlering in reaction.js at atom/guimaker/interfaces. Completed on 2024-09-11
 - Apply querySelectorAll instead querySelector only in reaction.js at atom/guimaker/interfaces where are the current method will miss addlistener to different element but with same feature such as same set of class. Completed on 2024-09-11
+- Add keyboard event to guimaker/browser/utils. Completed on 2024-09-13
 
 ### Changed